feat(Chat): move package chat to matchpost
