+++ src/main/java/com/sequence/anonymous/matchpost/domain/repository/ChatRepository.java
@@ -1,6 +1,6 @@
-package com.sequence.anonymous.chat.domain.repository;
+package com.sequence.anonymous.matchpost.domain.repository;
 
-import com.sequence.anonymous.chat.domain.Chat;
+import com.sequence.anonymous.matchpost.domain.Chat;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
