+++ src/main/java/com/sequence/anonymous/matchpost/presentation/ChatController.java
@@ -1,6 +1,6 @@
-package com.sequence.anonymous.chat.presentation;
+package com.sequence.anonymous.matchpost.presentation;
 
-import com.sequence.anonymous.chat.application.ChatService;
+import com.sequence.anonymous.matchpost.application.ChatService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
